+++ web/src/features/hfiCalculator/components/stationAdmin/AdminStationDropdown.tsx
@@ -1,6 +1,6 @@
 import { TextField, Autocomplete } from '@mui/material'
 import makeStyles from '@mui/styles/makeStyles'
-import { BasicWFWXStation, StationAdminRow } from 'features/hfiCalculator/components/stationAdmin/AddStationModal'
+import { BasicWFWXStation, StationAdminRow } from 'features/hfiCalculator/components/stationAdmin/ManageStationsModal'
 import { isEqual, isNull, isUndefined } from 'lodash'
 import React from 'react'
