+++ web/src/features/hfiCalculator/components/stationAdmin/PlanningAreaAdmin.tsx
@@ -3,7 +3,7 @@ import { PlanningArea } from 'api/hfiCalculatorAPI'
 import { Typography, Box, IconButton } from '@mui/material'
 import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline'
 import ExistingStationList from 'features/hfiCalculator/components/stationAdmin/ExistingStationList'
-import { AddStationOptions, StationAdminRow } from 'features/hfiCalculator/components/stationAdmin/AddStationModal'
+import { AddStationOptions, StationAdminRow } from 'features/hfiCalculator/components/stationAdmin/ManageStationsModal'
 import { AdminHandlers } from 'features/hfiCalculator/components/stationAdmin/StationListAdmin'
 import NewStationList from 'features/hfiCalculator/components/stationAdmin/NewStationList'
 import { differenceWith, isEqual } from 'lodash'
