+++ web/src/features/hfiCalculator/components/stationAdmin/adminFuelTypeDropdown.test.tsx
@@ -1,7 +1,7 @@
 import { render, waitFor, within } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { FuelType } from 'api/hfiCalculatorAPI'
-import { StationAdminRow } from 'features/hfiCalculator/components/stationAdmin/AddStationModal'
+import { StationAdminRow } from 'features/hfiCalculator/components/stationAdmin/ManageStationsModal'
 import { AdminFuelTypesDropdown } from 'features/hfiCalculator/components/stationAdmin/AdminFuelTypesDropdown'
 import React from 'react'
