+++ web/src/features/hfiCalculator/components/stationAdmin/manageStationsButton.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, waitFor } from '@testing-library/react'
 import store from 'app/store'
-import AddStationButton from 'features/hfiCalculator/components/stationAdmin/AddStationButton'
+import ManageStationsButton from 'features/hfiCalculator/components/stationAdmin/ManageStationsButton'
 import React from 'react'
 import { Provider } from 'react-redux'
 
@@ -8,7 +8,7 @@ describe('ManageStationsButton', () => {
   it('should render the button without the modal showing', () => {
     const { getByTestId, queryByText } = render(
       <Provider store={store}>
-        <AddStationButton planningAreaStationInfo={{}} />
+        <ManageStationsButton planningAreaStationInfo={{}} />
       </Provider>
     )
 
@@ -20,7 +20,7 @@ describe('ManageStationsButton', () => {
   it('should render the modal when the button is clicked', async () => {
     const { getByTestId } = render(
       <Provider store={store}>
-        <AddStationButton planningAreaStationInfo={{}} />
+        <ManageStationsButton planningAreaStationInfo={{}} />
       </Provider>
     )
